firebase buku

+++ app/src/main/java/com/example/perpustakaan/detailbuku/DetailActivity.kt
@@ -21,27 +21,27 @@ class DetailActivity : AppCompatActivity() {
 
     private lateinit var btnEdit: Button
     private lateinit var btnHapus: Button
-    private var id: Long? = null // Change to Long to match getLongExtra
+    private var id: Int? = null // Mengubah id menjadi Int
     private lateinit var judulBuku: TextView
     private lateinit var penulisBuku: TextView
     private lateinit var tahunBuku: TextView
     private lateinit var deskripsiBuku: TextView
-    private lateinit var imageBuku: ImageView // ImageView for book image
+    private lateinit var imageBuku: ImageView // ImageView untuk gambar buku
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.detailbuku)
 
-        // Retrieve data from Intent
-        id = intent.getLongExtra("BUKU_ID", -1L) // Use Long for ID
+        // Mengambil data dari Intent
+        id = intent.getIntExtra("BUKU_ID", -1) // Menggunakan Int untuk ID (default -1 untuk ID tidak valid)
 
-        if (id == -1L) { // Compare with -1L, not null
+        if (id == -1) { // Memeriksa apakah ID valid
             Toast.makeText(this, "Buku tidak ditemukan", Toast.LENGTH_SHORT).show()
-            finish()
+            finish() // Menutup activity jika buku tidak ditemukan
             return
         }
 
-        // Initialize views
+        // Inisialisasi tampilan
         btnEdit = findViewById(R.id.btn_edit_buku)
         btnHapus = findViewById(R.id.btn_hapus_buku)
         judulBuku = findViewById(R.id.tv_judul_buku)
@@ -50,37 +50,39 @@ class DetailActivity : AppCompatActivity() {
         deskripsiBuku = findViewById(R.id.tv_deskripsi)
         imageBuku = findViewById(R.id.img_buku)
 
-        // Load book details
+        // Memuat detail buku
         loadBukuDetails()
 
+        // Listener tombol Edit
         btnEdit.setOnClickListener {
             val intent = Intent(this, EditDataBukuActivity::class.java).apply {
-                putExtra("BUKU_ID", id)
-                putExtra("JUDUL", judulBuku.text.toString())
+                putExtra("BUKU_ID", id)  // Mengirimkan ID buku
+                putExtra("JUDUL", judulBuku.text.toString())  // Mengirimkan detail buku
                 putExtra("PENULIS", penulisBuku.text.toString())
                 putExtra("TAHUN_TERBIT", tahunBuku.text.toString())
                 putExtra("DESKRIPSI", deskripsiBuku.text.toString())
-                putExtra("GAMBAR_URL", imageBuku.tag?.toString()) // Ensure image URL tag is not null
+                putExtra("GAMBAR_URL", imageBuku.tag?.toString())  // Mengirimkan URL gambar dari tag ImageView
             }
             startActivity(intent)
         }
 
+        // Listener tombol Hapus
         btnHapus.setOnClickListener {
             lifecycleScope.launch(Dispatchers.IO) {
                 id?.let {
                     val daoBuku = PerpustakaanDatabase.getDatabase(this@DetailActivity).daobuku()
 
-                    // Ambil URL gambar dari ImageView atau database
-                    val gambarUrl = imageBuku.tag?.toString() // Gambar URL disimpan di tag ImageView
+                    // Mengambil URL gambar dari tag ImageView (mungkin disimpan dalam tag ImageView)
+                    val gambarUrl = imageBuku.tag?.toString()
 
-                    // Hapus buku berdasarkan ID dari database
+                    // Menghapus buku dari database berdasarkan ID
                     daoBuku.deleteBukuById(it)
 
-                    // Jika gambar ada dan URL-nya valid, hapus file gambar di cache
+                    // Jika ada URL gambar, mencoba menghapus file gambar dari penyimpanan lokal
                     gambarUrl?.let { url ->
-                        val file = File(url)  // Membuat File objek berdasarkan path gambar
+                        val file = File(url) // Membuat objek File berdasarkan path URL gambar
                         if (file.exists()) {
-                            val deleted = file.delete()  // Menghapus file gambar dari cache
+                            val deleted = file.delete()  // Mencoba menghapus file gambar dari cache
                             if (deleted) {
                                 Log.d("DetailActivity", "Gambar berhasil dihapus: $url")
                             } else {
@@ -90,9 +92,8 @@ class DetailActivity : AppCompatActivity() {
                     }
 
                     runOnUiThread {
-                        // Tampilkan toast jika buku berhasil dihapus
                         Toast.makeText(this@DetailActivity, "Buku dan gambarnya dihapus", Toast.LENGTH_SHORT).show()
-                        finish()  // Tutup activity setelah penghapusan
+                        finish()  // Menutup activity setelah penghapusan
                     }
                 }
             }
@@ -102,37 +103,37 @@ class DetailActivity : AppCompatActivity() {
     private fun loadBukuDetails() {
         id?.let {
             lifecycleScope.launch(Dispatchers.IO) {
-                // Debug log to see the book ID being queried
+                // Log untuk memeriksa apakah ID buku yang benar sedang diproses
                 Log.d("DetailActivity", "Memulai query dengan ID: $it")
 
                 val buku = PerpustakaanDatabase.getDatabase(this@DetailActivity)
                     .daobuku()
-                    .getBukuById(it) // Query database based on ID
+                    .getBukuById(it)  // Mengambil data buku berdasarkan ID
 
                 if (buku != null) {
                     runOnUiThread {
-                        // Display book details on the UI
+                        // Memperbarui elemen UI dengan detail buku dari database
                         judulBuku.text = buku.judul
                         penulisBuku.text = buku.penulis
                         tahunBuku.text = buku.tahunTerbit.toString()
                         deskripsiBuku.text = buku.deskripsi
 
-                        // Save image URL in the tag of ImageView for future use
+                        // Menyimpan URL gambar di tag ImageView
                         imageBuku.tag = buku.gambarUrl
 
-                        // Load image using Glide
+                        // Memuat gambar dari URL menggunakan Glide
                         Glide.with(this@DetailActivity)
-                            .load(buku.gambarUrl) // Image URL from database
-                            .into(imageBuku)
+                            .load(buku.gambarUrl)  // Menggunakan URL gambar yang diambil dari database
+                            .into(imageBuku)  // Menampilkan gambar di ImageView
                     }
                 } else {
-                    // Debug log if book is not found
+                    // Log jika buku tidak ditemukan
                     Log.e("DetailActivity", "Buku tidak ditemukan dengan ID: $it")
 
                     runOnUiThread {
-                        // If book not found, show a message
+                        // Jika buku tidak ditemukan, tampilkan pesan
                         Toast.makeText(this@DetailActivity, "Buku tidak ditemukan", Toast.LENGTH_SHORT).show()
-                        finish()
+                        finish()  // Menutup activity jika buku tidak ditemukan
                     }
                 }
             }
